+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Buffer completion now offers only the current project's buffers
   unless the command using it was called with a prefix argument.
+- Command `ship-mate-show-results` to pop to a buffer. Bound to `p` in
+  subcommands map.
 
 ### Changed